Refactor the QueryCompiler module to separate backends
* Preparing to factor out Metadata management into a separate class
* Create `modin.backends.<base, pandas, pyarrow>
