--- modin/data_management/query_compiler/__init__.py
@@ -1,11 +1,0 @@
-from .pandas_query_compiler import PandasQueryCompiler, PandasQueryCompilerView
-from .base_query_compiler import BaseQueryCompiler, BaseQueryCompilerView
-from .pyarrow_query_compiler import PyarrowQueryCompiler
-
-__all__ = [
-    "PandasQueryCompiler",
-    "PandasQueryCompilerView",
-    "BaseQueryCompiler",
-    "BaseQueryCompilerView",
-    "PyarrowQueryCompiler",
-]
