+++ modin/engines/python/pandas_on_python/io.py
@@ -3,7 +3,7 @@ from __future__ import division
 from __future__ import print_function
 
 from modin.engines.base.io import BaseIO
-from modin.data_management.query_compiler import PandasQueryCompiler
+from modin.backends.pandas.query_compiler import PandasQueryCompiler
 from modin.engines.python.pandas_on_python.frame.partition_manager import (
     PythonFrameManager,
 )
