+++ modin/engines/ray/pandas_on_ray/io.py
@@ -7,7 +7,7 @@ import pandas
 from io import BytesIO
 import ray
 from modin.data_management.utils import split_result_of_axis_func_pandas
-from modin.data_management.query_compiler import PandasQueryCompiler
+from modin.backends.pandas.query_compiler import PandasQueryCompiler
 from modin.engines.ray.generic.io import RayIO
 from modin.engines.ray.pandas_on_ray.frame.partition_manager import (
     PandasOnRayFrameManager,
